+++ plugins/rbac/CHANGELOG.md
@@ -1,3 +1,16 @@
+## @janus-idp/backstage-plugin-rbac [1.24.1](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-rbac@1.24.0...@janus-idp/backstage-plugin-rbac@1.24.1) (2024-07-24)
+
+
+### Bug Fixes
+
+* **deps:** rollback unreleased plugins ([#1951](https://github.com/janus-idp/backstage-plugins/issues/1951)) ([8b77969](https://github.com/janus-idp/backstage-plugins/commit/8b779694f02f8125587296305276b84cdfeeaebe))
+
+
+
+### Dependencies
+
+* **@janus-idp/backstage-plugin-rbac-common:** upgraded to 1.7.2
+
 ## @janus-idp/backstage-plugin-rbac [1.24.0](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-rbac@1.23.2...@janus-idp/backstage-plugin-rbac@1.24.0) (2024-07-24)